+++ modules/audit_system_auth_unlock_time.sh
@@ -19,39 +19,39 @@ audit_system_auth_unlock_time () {
   search_string=$2
   search_value=$3
   if [ "$os_name" = "Linux" ]; then
-    if [ "$os_vendor" = "Red" ] || [ "$os_vendor" = "CentOS" ]; then
-      check_file="/etc/pam.d/system-auth"
-    fi
-    if [ "$os_vendor" = "Debian" ] || [ "$os_vendor" = "SuSE" ] || [ "$os_vendor" = "Ubuntu" ]; then
-      check_file="/etc/pam.d/common-auth"
-    fi
-    if [ "$audit_mode" != 2 ]; then
-      echo "Checking:  Lockout time for failed password attempts enabled in $check_file"
-      total=`expr $total + 1`
-      check_value=`cat $check_file |grep '^$auth_string' |grep '$search_string$' |awk -F '$search_string=' '{print $2}' |awk '{print $1}'`
-      if [ "$check_value" != "$search_string" ]; then
-        if [ "$audit_mode" = "1" ]; then
-          insecure=`expr $insecure + 1`
-          echo "Warning:   Lockout time for failed password attempts not enabled in $check_file [$insecure Warnings]"
-          funct_verbose_message "cp $check_file $temp_file" fix
-          funct_verbose_message "cat $temp_file |sed 's/^auth.*pam_env.so$/&\nauth\t\trequired\t\t\tpam_faillock.so preauth audit silent deny=5 unlock_time=900\nauth\t\t[success=1 default=bad]\t\t\tpam_unix.so\nauth\t\t[default=die]\t\t\tpam_faillock.so authfail audit deny=5 unlock_time=900\nauth\t\tsufficient\t\t\tpam_faillock.so authsucc audit deny=5 $search_string=$search_value\n/' > $check_file" fix
-          funct_verbose_message "rm $temp_file" fix
-        fi
-        if [ "$audit_mode" = 0 ]; then
-          funct_backup_file $check_file
-          echo "Setting:   Password minimum length in $check_file"
-          cp $check_file $temp_file
-          cat $temp_file |sed 's/^auth.*pam_env.so$/&\nauth\t\trequired\t\t\tpam_faillock.so preauth audit silent deny=5 unlock_time=900\nauth\t\t[success=1 default=bad]\t\t\tpam_unix.so\nauth\t\t[default=die]\t\t\tpam_faillock.so authfail audit deny=5 unlock_time=900\nauth\t\tsufficient\t\t\tpam_faillock.so authsucc audit deny=5 $search_string=$search_value\n/' > $check_file
-          rm $temp_file
-        fi
-      else
-        if [ "$audit_mode" = "1" ]; then
-          secure=`expr $secure + 1`
-          echo "Secure:    Lockout time for failed password attempts enabled in $check_file [$secure Passes]"
-        fi
+    for check_file in /etc/pam.d/system-auth /etc/pam.d/common-auth; do
+      if [ -f "$check_file" ]; then
+        if [ "$audit_mode" != 2 ]; then
+          echo "Checking:  Lockout time for failed password attempts enabled in $check_file"
+          total=`expr $total + 1`
+          check_value=`cat $check_file |grep '^$auth_string' |grep '$search_string$' |awk -F '$search_string=' '{print $2}' |awk '{print $1}'`
+          if [ "$check_value" != "$search_string" ]; then
+            if [ "$audit_mode" = "1" ]; then
+              insecure=`expr $insecure + 1`
+              echo "Warning:   Lockout time for failed password attempts not enabled in $check_file [$insecure Warnings]"
+              funct_verbose_message "cp $check_file $temp_file" fix
+              funct_verbose_message "cat $temp_file |sed 's/^auth.*pam_env.so$/&\nauth\t\trequired\t\t\tpam_faillock.so preauth audit silent deny=5 unlock_time=900\nauth\t\t[success=1 default=bad]\t\t\tpam_unix.so\nauth\t\t[default=die]\t\t\tpam_faillock.so authfail audit deny=5 unlock_time=900\nauth\t\tsufficient\t\t\tpam_faillock.so authsucc audit deny=5 $search_string=$search_value\n/' > $check_file" fix
+              funct_verbose_message "rm $temp_file" fix
+            fi
+            if [ "$audit_mode" = 0 ]; then
+              funct_backup_file $check_file
+              echo "Setting:   Password minimum length in $check_file"
+              cp $check_file $temp_file
+              cat $temp_file |sed 's/^auth.*pam_env.so$/&\nauth\t\trequired\t\t\tpam_faillock.so preauth audit silent deny=5 unlock_time=900\nauth\t\t[success=1 default=bad]\t\t\tpam_unix.so\nauth\t\t[default=die]\t\t\tpam_faillock.so authfail audit deny=5 unlock_time=900\nauth\t\tsufficient\t\t\tpam_faillock.so authsucc audit deny=5 $search_string=$search_value\n/' > $check_file
+              rm $temp_file
+            fi
+          else
+            if [ "$audit_mode" = "1" ]; then
+              secure=`expr $secure + 1`
+              echo "Secure:    Lockout time for failed password attempts enabled in $check_file [$secure Passes]"
+            fi
+          fi
+        done
       fi
     else
-      funct_restore_file $check_file $restore_dir
+      for check_file in /etc/pam.d/system-auth /etc/pam.d/common-auth; do
+        funct_restore_file $check_file $restore_dir
+      done
     fi
   fi
 }
